refactor: dialog refactor
