+++ storybook/lib/main.dart
@@ -9,14 +9,14 @@ import 'package:storybook/stories/button/split.dart';
 import 'package:storybook/stories/card.dart';
 import 'package:storybook/stories/checkbox.dart';
 import 'package:storybook/stories/checkbox_group.dart';
-import 'package:storybook/stories/dialog.dart';
+import 'package:storybook/stories/dialog/dialog.dart';
 import 'package:storybook/stories/form/form_story.dart';
 import 'package:storybook/stories/icon.dart';
 import 'package:storybook/stories/icon_list.dart';
 import 'package:storybook/stories/input.dart';
 import 'package:storybook/stories/list_tile.dart';
 import 'package:storybook/stories/nested_overlays.dart';
-import 'package:storybook/stories/nonmodal_wrapper.dart';
+import 'package:storybook/stories/dialog/non_modal_wrapper.dart';
 import 'package:storybook/stories/number_picker.dart';
 import 'package:storybook/stories/radio.dart';
 import 'package:storybook/stories/search_field.dart';
